Inserting CustmoerBuilder inside Customer class
